// src/utils/enums.ts
export enum UserStatus {
  ACTIVE = "ACTIVE",
  INACTIVE = "INACTIVE",
  BANED = "BANED",
}

export enum UserRole {
  USER = "USER",
  TUTOR = "TUTOR",
  ADMIN = "ADMIN",
  ALL = "ALL",
}

export enum TutorStatus {
  APPROVED = "APPROVED",
  PENDING = "PENDING",
}

export enum TutorActivity {
  ACTIVE = "ACTIVE",
  INACTIVE = "INACTIVE",
  BANNED = "BANNED",
}